+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * AbcdSeq: added menu options to copy/paste each row individually
   * AbcdSeq: added button/input to randomize all length knobs
   * AbcdSeq: added menu option to use velocity as a probability
+  * AbcdSeq: added end of cycle per row
   * Arrange: colored text fields and input/output ports
 
 ## v2.0.14 ~ 5-29-25